feat(etcd-manager-ctl): use backupname to delete backup instead of timestamp
This commit refactors the `runDeleteCommand` function of
`etcd-manager-ctl` to use the "backup name" instead of the "timestamp"
attribute. More often than not, users are more likely to become familiar
with the syntax of `etcd-manager-ctl restore-backup`, which uses the
`backupname` attribute to identify the target backup to restore.

By bringing the `delete-backup` subcommand in alignment with
`restore-backup`, we improve the user experience of an operator in the
process of restoring a cluster.