Build base image by raw expansion of deb packages
This is how we used to build our images before bazel discouraged it,
but is similar to how distroless is built now.  It has many fewer
dependencies compared to the distroless stack, and is much more
reproducible than Dockerfile alternatives.  This is currently
integrated to bazel, but we could also do this standalone or in Docker
in future.